20240923 backup

+++ src/main/java/hcnc/cteam/attendance/AttenDTO.java
@@ -13,7 +13,7 @@ public class AttenDTO {
 	private String attenType; // 근무 형태
 	private LocalTime workStart; // 출근 시간
 	private LocalTime workEnd; // 퇴근 시간
-	private long workOver; // 초과 근무 시간
+	private double workOver; // 초과 근무 시간
 	
 	public int getEmpCode() {
 		return empCode;
@@ -63,10 +63,10 @@ public class AttenDTO {
 	public void setWorkEnd(LocalTime workEnd) {
 		this.workEnd = workEnd;
 	}
-	public long getWorkOver() {
+	public double getWorkOver() {
 		return workOver;
 	}
-	public void setWorkOver(long workOver) {
+	public void setWorkOver(double workOver) {
 		this.workOver = workOver;
 	}
